// apps/wallet/types/customerPaymentData.ts
export interface CustomerPaymentData {
  createdAt: string;
  customerId: string;
  updatedAt: string;
  accountId: string;
  type: string;
  amount: string;
  usd: string;
  currency: string;
  poolId: string;
  tokenId: string;
  tokenQuantity: string;
  transactionId: string;
  cancelTransactionId: string;
  status: string;
  recordId: string;
  processedRecordId: string;
  cancelRecordId: string;
  observations: string;
}